Added Rhythm Game
Added rhythm game to the main.py file, I had to add JavaScrip to the bottom of the json file to fetch the data from generate_rhytm, and added a script source to the index.html ((now called rhtyhm_index.html)). I ran it through my local host and flask and it worked, so I am pushing my changes now and then pushing the changes to heroku.