Updating some style components.
I changed the overall style of the website to go along with the mine craft-esque theme. It is kind of ugly right now but I wanted to see what it would look like. I am pushing it to the main so others can see what it looks like and I believe if they don't see the most updated version it could hinder their progress potentially.