working again on the sound game
I tried a new implemenatation of the sound game files by copy and pasting them into the main.py file ad then tweaking the names so the same methods were not being overwritten by the two modes. This is the closest I have come to successfully implementing this game on the site.